Move the Parse effect into its own module.
